+++ lizmap/test/test_table_checks.py
@@ -2,6 +2,10 @@ import unittest
 
 from lizmap.widgets.check_project import Checks, Error, TableCheck
 
+__copyright__ = 'Copyright 2023, 3Liz'
+__license__ = 'GPL version 3'
+__email__ = 'info@3liz.org'
+
 
 class TestProjectTable(unittest.TestCase):